Move everything into KaizenDraft dir
